requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
fake-useragent==2.2.0
idna==3.10
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
